Added labels to all BQ queries, refactoring billing layer.
