+++ db/python/layers/billing_db.py
@@ -15,7 +15,6 @@ from api.settings import (
 )
 from api.utils.dates import get_invoice_month_range, reformat_datetime
 from db.python.gcp_connect import BqDbBase
-from db.python.layers.bq_base import BqBaseLayer
 from db.python.tables.billing import BillingFilter
 from models.models import (
     BillingColumn,
@@ -25,157 +24,12 @@ from models.models import (
     BillingTotalCostRecord,
 )
 
+BQ_LABELS = {'source': 'metamist-api'}
+
 
 def abbrev_cost_category(cost_category: str) -> str:
     """abbreviate cost category"""
     return 'S' if cost_category == 'Cloud Storage' else 'C'
-
-
-class BillingLayer(BqBaseLayer):
-    """Billing layer"""
-
-    async def get_gcp_projects(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All GCP projects in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_gcp_projects()
-
-    async def get_topics(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All topics in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_topics()
-
-    async def get_cost_categories(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All service description / cost categories in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_cost_categories()
-
-    async def get_skus(
-        self,
-        limit: int | None = None,
-        offset: int | None = None,
-    ) -> list[str] | None:
-        """
-        Get All SKUs in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_skus(limit, offset)
-
-    async def get_datasets(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All datasets in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('dataset')
-
-    async def get_stages(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All stages in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('stage')
-
-    async def get_sequencing_types(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All sequencing_types in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('sequencing_type')
-
-    async def get_sequencing_groups(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All sequencing_groups in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('sequencing_group')
-
-    async def get_compute_categories(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All compute_category values in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('compute_category')
-
-    async def get_cromwell_sub_workflow_names(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All cromwell_sub_workflow_name values in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('cromwell_sub_workflow_name')
-
-    async def get_wdl_task_names(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All wdl_task_name values in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_extended_values('wdl_task_name')
-
-    async def get_invoice_months(
-        self,
-    ) -> list[str] | None:
-        """
-        Get All invoice months in database
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_invoice_months()
-
-    async def query(
-        self,
-        _filter: BillingFilter,
-        limit: int = 10,
-    ) -> list[BillingRowRecord] | None:
-        """
-        Get Billing record for the given gilter
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.query(_filter, limit)
-
-    async def get_total_cost(
-        self,
-        query: BillingTotalCostQueryModel,
-    ) -> list[BillingTotalCostRecord] | None:
-        """
-        Get Total cost of selected fields for requested time interval
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_total_cost(query)
-
-    async def get_running_cost(
-        self,
-        field: BillingColumn,
-        invoice_month: str | None = None,
-        source: str | None = None,
-    ) -> list[BillingCostBudgetRecord]:
-        """
-        Get Running costs including monthly budget
-        """
-        billing_db = BillingDb(self.connection)
-        return await billing_db.get_running_cost(field, invoice_month, source)
 
 
 class BillingDb(BqDbBase):
@@ -203,7 +57,8 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter(
                     'days', 'INT64', -int(BQ_DAYS_BACK_OPTIMAL)
                 ),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -237,7 +92,8 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter(
                     'days', 'INT64', -int(BQ_DAYS_BACK_OPTIMAL)
                 ),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -259,7 +115,11 @@ class BillingDb(BqDbBase):
         ORDER BY invoice_month DESC;
         """
 
-        query_job_result = list(self._connection.connection.query(_query).result())
+        job_config = bigquery.QueryJobConfig(labels=BQ_LABELS)
+
+        query_job_result = list(
+            self._connection.connection.query(_query, job_config=job_config).result()
+        )
         if query_job_result:
             return [str(dict(row)['invoice_month']) for row in query_job_result]
 
@@ -288,7 +148,8 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter(
                     'days', 'INT64', -int(BQ_DAYS_BACK_OPTIMAL)
                 ),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -334,7 +195,8 @@ class BillingDb(BqDbBase):
                 ),
                 bigquery.ScalarQueryParameter('limit_val', 'INT64', limit),
                 bigquery.ScalarQueryParameter('offset_val', 'INT64', offset),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -375,7 +237,8 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter(
                     'days', 'INT64', -int(BQ_DAYS_BACK_OPTIMAL)
                 ),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -446,7 +309,9 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter('limit_val', 'INT64', limit)
             )
 
-        job_config = bigquery.QueryJobConfig(query_parameters=query_parameters)
+        job_config = bigquery.QueryJobConfig(
+            query_parameters=query_parameters, labels=BQ_LABELS
+        )
         query_job_result = list(
             self._connection.connection.query(_query, job_config=job_config).result()
         )
@@ -559,7 +424,9 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter('offset_val', 'INT64', query.offset)
             )
 
-        job_config = bigquery.QueryJobConfig(query_parameters=query_parameters)
+        job_config = bigquery.QueryJobConfig(
+            query_parameters=query_parameters, labels=BQ_LABELS
+        )
         query_job_result = list(
             self._connection.connection.query(_query, job_config=job_config).result()
         )
@@ -593,7 +460,10 @@ class BillingDb(BqDbBase):
         ON d.gcp_project = t.gcp_project AND d.created_at = t.last_created_at
         """
 
-        query_job_result = list(self._connection.connection.query(_query).result())
+        job_config = bigquery.QueryJobConfig(labels=BQ_LABELS)
+        query_job_result = list(
+            self._connection.connection.query(_query, job_config=job_config).result()
+        )
 
         if query_job_result:
             return {row.gcp_project: row.budget for row in query_job_result}
@@ -619,7 +489,8 @@ class BillingDb(BqDbBase):
                 bigquery.ScalarQueryParameter(
                     'days', 'INT64', -int(BQ_DAYS_BACK_OPTIMAL)
                 ),
-            ]
+            ],
+            labels=BQ_LABELS,
         )
 
         query_job_result = list(
@@ -792,7 +663,9 @@ class BillingDb(BqDbBase):
             list(
                 self._connection.connection.query(
                     _query,
-                    job_config=bigquery.QueryJobConfig(query_parameters=query_params),
+                    job_config=bigquery.QueryJobConfig(
+                        query_parameters=query_params, labels=BQ_LABELS
+                    ),
                 ).result()
             ),
         )
